Rule converter now as stand alone!
Conversion can be started now via terminal. Rule Converter needs some parameters to execute the conversion successfully.
-o file path to output
-s file path for scheme
-p file path for pmml

Therefore JCommander is isued and included in jar file. In addition the converter class can be used for integrated solutions, of course.
To use it, you will need to instantiate a converter object with filepath to converter schema (dtd have to be found at the same place).
After that the convert() function can be used. For that a list of pmml paths is necessary. After conversion this function will return a string.
See usage also in ConverterMain class and examples in rule-convert/examples.